+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 - 予約取消アクションのIObjectを最適化
 - イベントオファー承認アクションのobject.eventを最適化
+- 予約のIBrokerを最適化
 
 ### Deprecated